Intent: Reorganize canonical grammar validation flows Operators involved: [Transition|Mutation|Coherence] Affected invariants: [#4, #5, #7] Key changes: - Move GrammarContext and grammar enforcement helpers into tnfr.validation.grammar to isolate canonical enforcement context - Split canonical compatibility tables into tnfr.validation.compatibility and validation rules into tnfr.validation.rules with focused helpers - Update imports and add validation tests covering grammar, compatibility, and rules modules to preserve TNFR invariants Expected risks/dissonances: Potential third-party imports still referencing tnfr.grammar; mitigated with compatibility re-exports and targeted tests Metrics: Expect unchanged C(t), Si, νf; normalisation thresholds exercised in new tests Equivalence map: tnfr.grammar.* -> tnfr.validation.grammar.*, tnfr.grammar.CANON_* -> tnfr.validation.compatibility.CANON_*
