+++ src/tnfr/validation/__init__.py
@@ -1,5 +1,25 @@
 """Validation helpers for TNFR sequences."""
 
+from .compatibility import CANON_COMPAT, CANON_FALLBACK
+from .grammar import (
+    GrammarContext,
+    apply_glyph_with_grammar,
+    enforce_canonical_grammar,
+    on_applied_glyph,
+)
+from .rules import coerce_glyph, glyph_fallback, normalized_dnfr, get_norm
 from .syntax import validate_sequence
 
-__all__ = ("validate_sequence",)
+__all__ = (
+    "validate_sequence",
+    "GrammarContext",
+    "apply_glyph_with_grammar",
+    "enforce_canonical_grammar",
+    "on_applied_glyph",
+    "coerce_glyph",
+    "glyph_fallback",
+    "normalized_dnfr",
+    "get_norm",
+    "CANON_COMPAT",
+    "CANON_FALLBACK",
+)
